requirements.txt
youtube_transcript_api
anthropic
supabase
requests
json
numpy
google.generativeai
pinecone
fastapi
